refactor: extract construction contract normalizers
