+++ make/ci.mk
@@ -716,6 +716,7 @@ ci.local.quick: guard.prod.forbid architecture.complexity_baseline_lock verify.u
 	@python3 scripts/verify/construction_core_extension_capability_rows_split_guard.py
 	@python3 scripts/verify/construction_core_extension_hook_facts_split_guard.py
 	@python3 scripts/verify/construction_core_extension_policy_accessors_split_guard.py
+	@python3 scripts/verify/construction_core_extension_contract_normalizers_split_guard.py
 	@python3 scripts/verify/construction_core_extension_responsibility_map_guard.py
 	@python3 scripts/verify/frontend_page_contract_boundary_guard.py
 	@python3 scripts/verify/frontend_page_contract_orchestration_consumption_guard.py
